Serialize test access to clipboard.
The problem this solves is that when parallel testing is requested the
various individual tests were failing because they were seeing each
other's clipboard modifications.

Since the test files were written to be independant, the simplest way to
serialize them under parallel testing seemed to be to rename them so
that the test harness would no longer see them as tests and then write
an umbrella test that used 'do' to execute each one in the desired
order.

This involves an increase in the required version of Test::More, to
0.96.