Tweak t/pasteboard.t do() file names.
This test does things like "do 't/misc.tx';", and some CPAN testers
flunk this with the error "... '.' is no longer in @INC ...". So I
accepted the error message's recommendation and made it './t/misc.tx'.

Oddly, this does not fail on my box, for the same version of Perl.