Stopgap mods for macOS 10.15 Catalina
The underlying problem appears to be that copy() does not put stuff on
the public pasteboards. It puts it somewhere, though, becase paste() can
find it, even if run in a separate process. But paste() also finds stuff
put on the pasteboard by pbcopy (1).

The above makes no sense to me, but I band-aided it by having copy()
spawn pbcopy if POSIX::uname indicates that we are running under Darwin
19 or higher.