Replace SvPV() with SvPVbyte().
This was triggered by Felipe Gasper's blog entry "Perl's SvPV Menace" at
https://dev.to/fgasper/perl-s-svpv-menace-5515

Of course, it's not that simple because Perl 5.6 does not have SvPVbyte.
So I ended up with macros my_SvPVbyte().

But it's not even that simple, because pasteboard and flavor names are
also strings, and I have yet to deal with them. Maybe I need to use
SvPVutf8() there?